+++ scripts/experiments/symmetries/src/test_block_decomposition.py
@@ -1,8 +1,9 @@
 import numpy as np
-from block_decomposition import matmul_mod, modp, is_symplectic, block_decompose, rank_mod, inv_mod_mat
+from block_decomposition import matmul_mod, modp, is_symplectic, block_decompose, ordered_block_sizes, inv_mod_mat
 # =========================
 # Generators & test: scrambled SWAP
 # =========================
+
 
 def swap_symplectic(n: int, i: int, j: int, p: int) -> np.ndarray:
     """Swap qudits i and j (0-based) in phase space: diag(P, P) in [x|z] ordering."""
@@ -12,6 +13,7 @@ def swap_symplectic(n: int, i: int, j: int, p: int) -> np.ndarray:
     O = np.zeros((n, n), dtype=np.int64)
     return np.block([[P, O],
                      [O, P]]) % p
+
 
 def random_symplectic(n: int, p: int, rng=None, steps: int = 5) -> np.ndarray:
     """Lightweight scrambler (not uniform) built from generators preserving Ω."""
@@ -45,6 +47,7 @@ def random_symplectic(n: int, p: int, rng=None, steps: int = 5) -> np.ndarray:
         F = matmul_mod(L, F, p)
     assert is_symplectic(F, p)
     return F
+
 
 def symplectic_block_sizes_from_mode_graph(S: np.ndarray, p: int) -> list[int]:
     """
@@ -92,56 +95,6 @@ def symplectic_block_sizes_from_mode_graph(S: np.ndarray, p: int) -> list[int]:
     return sizes
 
 
-def ordered_block_sizes(S: np.ndarray, p: int) -> list[int]:
-    n2 = S.shape[0]
-    k = n2 // 2
-    # compute rank of (S - I) restricted to each component as nontriviality
-    M = (S - np.eye(n2, dtype=np.int64)) % p
-
-    # build comps first (copy the same adjacency/components code)
-    adj = [[] for _ in range(k)]
-    for i in range(k):
-        rows_i = [i, k + i]
-        for j in range(i + 1, k):
-            cols_j = [j, k + j]
-            if (S[np.ix_(rows_i, cols_j)] % p).any() or (S[np.ix_([j, k + j], [i, k + i])] % p).any():
-                adj[i].append(j)
-                adj[j].append(i)
-    seen = [False]*k
-    comps = []
-    for s in range(k):
-        if seen[s]:
-            continue
-        st = [s]
-        seen[s] = True
-        comp = []
-        while st:
-            u = st.pop()
-            comp.append(u)
-            for v in adj[u]:
-                if not seen[v]:
-                    seen[v] = True
-                    st.append(v)
-        comps.append(sorted(comp))
-
-    def comp_rank(comp):
-        idx = comp + [c + k for c in comp]
-        # rank_mod must be your GF(p) rank
-        return rank_mod(M[np.ix_(idx, idx)], p)
-
-    info = [{"modes": c, "nmodes": len(c), "rank": comp_rank(c)} for c in comps]
-    nontriv_ge2 = [d for d in info if d["rank"] > 0 and d["nmodes"] >= 2]
-    nontriv_1 = [d for d in info if d["rank"] > 0 and d["nmodes"] == 1]
-    triv = [d for d in info if d["rank"] == 0]
-
-    nontriv_ge2.sort(key=lambda d: (d["nmodes"], -d["rank"]))
-    nontriv_1.sort(key=lambda d: -d["rank"])
-    triv.sort(key=lambda d: d["nmodes"])
-
-    ordered = nontriv_ge2 + nontriv_1 + triv
-    return [2 * d["nmodes"] for d in ordered]
-
-
 def test_scrambled_swap(p: int = 3, n: int = 6, i: int = 1, j: int = 4, seed: int = 42):
     rng = np.random.default_rng(seed)
     S_true = swap_symplectic(n, i, j, p)
@@ -164,5 +117,5 @@ def test_scrambled_swap(p: int = 3, n: int = 6, i: int = 1, j: int = 4, seed: in
 if __name__ == "__main__":
     for _ in range(100):
         seed = np.random.randint(0, 2**32)
-        S, T, F = test_scrambled_swap(p=17, n=20, i=1, j=2, seed=seed)
+        S, T, F = test_scrambled_swap(p=17, n=20, i=4, j=2, seed=seed)
     print('passed all tests')
